requirements.txt
python-dotenv~=1.1.0
requests~=2.32.4
PyYAML~=6.0.2
beautifulsoup4~=4.12.3
lxml>=5.1.0

# Dev dependencies
pytest~=8.0.0
pytest-mock~=3.12.0
ruff~=0.2.0
